requirements.txt
asn1crypto==0.24.0
CD==1.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cx-Oracle==7.3.0
DateTime==4.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-hitcount==1.3.0
django-import-export==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-tracking==1.5.0
et-xmlfile==1.0.1
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
mysql-connector-python==8.0.13
mysqlclient==1.3.14
numpy==1.18.5
odfpy==1.4.0
openpyxl==2.5.12
pandas==0.24.2
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
git+https://jmirapale:aCSMsNfh72VE4Sh9j38x@bitbucket.org/Emtelco_TIC/pyctivex.git
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.8.2
pytz==2018.6
PyYAML==3.13
requests==2.20.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
zope.interface==4.6.0
